server/requirements.txt
fastapi
uvicorn
requests
groq
python-dotenv
